refactor: reorganize folder structure inside of PolygonDraw
