+++ src/PolygonDraw/PolygonMap/components/BoundaryPolygon.tsx
@@ -2,8 +2,8 @@ import { FunctionComponent } from 'react';
 import { Polygon as LeafletPolygon } from 'react-leaflet';
 
 import { Coordinate } from 'types';
-import { createLeafletLatLngFromCoordinate } from '../helpers';
-import { MAP } from '../constants';
+import { createLeafletLatLngFromCoordinate } from '../../../helpers';
+import { MAP } from '../../../constants';
 
 interface Props {
     coordinates: Coordinate[];
