+++ src/PolygonDraw/PolygonMap/components/PolygonPane.tsx
@@ -1,15 +1,15 @@
 import { Pane } from 'react-leaflet';
 
-import { Coordinate } from '../../types';
+import { Coordinate } from '../../../types';
 import {
     addCoordinates,
     createCoordinateFromLeafletLatLng,
     getPolygonEdges,
     isCoordinateInPolygon,
     subtractCoordinates,
-} from '../../helpers';
-import { PolygonVertex } from '../PolygonVertex';
-import { EdgeVertex } from '../EdgeVertex';
+} from '../../../helpers';
+import { PolygonVertex } from './PolygonVertex';
+import { EdgeVertex } from './EdgeVertex';
 import { LatLng } from 'leaflet';
 import { useState } from 'react';
