+++ src/PolygonDraw/utils/validators.ts
@@ -1,5 +1,5 @@
 import { Coordinate } from 'types';
-import { isPolygonClosed } from '../helpers';
+import { isPolygonClosed } from '../../helpers';
 
 export const isValidPolygon = (coordinates: Coordinate[]): boolean =>
     !!coordinates && coordinates.length > 3 && isPolygonClosed(coordinates);
