+++ src/index.ts
@@ -1,2 +1,2 @@
 export { PolygonDraw } from './PolygonDraw/PolygonDraw';
-export { isValidPolygon } from './PolygonDraw/validators';
+export { isValidPolygon } from './PolygonDraw/utils/validators';
